\name{TCGAquery-package}
\alias{TCGAquery-package}
\alias{TCGAquery}
\docType{package}
\title{
Query and download data from the Cancer Genome Atlas (TCGA)
}
\description{
The Cancer Genome Atlas (TCGA http://cancergenome.nih.gov/) is an ongoing project on generating data from primary cancer samples and matched normals using many different genomic platforms. This package uses the TCGA web API to request data on any disease or platform, download the data archive and extract the data into R objects. Only support for public access data is provided at this time.
}
\details{
\tabular{ll}{
Package: \tab TCGAquery\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-09-16\cr
License: \tab LGPL-3\cr
}

}
\author{
Author: Sudhir Varma, HiThru Analytics LLC <TCGAquery@hithru.com>

Maintainer: Sudhir Varma, HiThru Analytics LLC <TCGAquery@hithru.com>
}
\references{

}

\keyword{ package }
\seealso{
  \code{
    \link[GEOquery:getGEO]{GEOquery}, \link[ArrayExpress:ArrayExpress]{ArrayExpress}
  }
}
\examples{
~~ simple examples of the most important functions ~~
}
